move construct dependency inference to the app level - and call it in chart.toJson() as well
